Fix BFloat16 numpy conversion in embedding extraction
Convert to float32 before numpy conversion since numpy doesn't support BFloat16.

+++ evals/run_detection.py
@@ -451,7 +451,7 @@ def get_letter_embeddings(model, tokenizer) -> Dict[str, np.ndarray]:
     else:
         raise ValueError("Unknown model architecture for embedding extraction")
 
-    embeddings = embed_layer.weight.detach().cpu().numpy()
+    embeddings = embed_layer.weight.detach().cpu().float().numpy()
 
     letter_embeddings = {}
     for letter in "abcdefghijklmnopqrstuvwxyz":
